Patch for error in messages by deleted users
when you go to projects->messages and there are messages posted by users
who have already been deleted, a fatal error is thrown.

To fix this, a condition has been put in place, to only try to display
the author of the message if the user is found in the database.